consumer/requirements.txt
confluent_kafka
boto3
python-dotenv
